requirements.txt
PyPDF2
pillow
starlette
uvicorn
loguru
python-multipart
pdf2image
google-generativeai
httpx